Removed comming updating the vendored sources (of libbuildpack).
Doing the update breaks the build, introducing a revision of
libbuildpack which does not match code using it. It is suspected that
the vendoring manifest refers to a different version than we have
actually vendored in the repo.

Generally speaking it should never be necessary to update vendored
code in a CI. Having the proper required revisions of dependencies in
the repository itself is what vendoring is for.